+++ src/Paddle.java
@@ -92,7 +92,7 @@ class Paddle implements GameObject{
      * Returns the y value of the top of the hitbox
      * @return the y value of the top edge of the paddle
      */
-    public int getTop() {
+    public double getTop() {
         return rectangle.getLayoutY() + rectangle.getTranslateY(); 
     }
      
@@ -100,7 +100,7 @@ class Paddle implements GameObject{
      * Returns the x value of the right of the hitbox
      * @return the x value of the right edge of the paddle
      */
-    public int getRight() {
+    public double getRight() {
         return rectangle.getLayoutX() + rectangle.getTranslateX() + PADDLE_WIDTH; 
     }   
 
@@ -108,7 +108,7 @@ class Paddle implements GameObject{
      * Returns the y value of the bottom of the hitbox
      * @return the y value of the bottom edge of the paddle
      */
-    public int getBottom() {
+    public double getBottom() {
         return rectangle.getLayoutY() + rectangle.getTranslateY() + PADDLE_HEIGHT; 
     }   
 
@@ -116,7 +116,7 @@ class Paddle implements GameObject{
      * Returns the x value of the left of the hitbox
      * @return the x value of the left edge of the paddle
      */
-    public int getLeft() {
+    public double getLeft() {
         return rectangle.getLayoutX() + rectangle.getTranslateX(); 
     }
 }
